backend/requirements.txt
# Core LangGraph and LangChain dependencies
torch
langgraph>=0.2.6
langchain>=0.3.19
langchain-openai
openai
duckduckgo-search>=5.0
python-dotenv>=1.0.1
langgraph-sdk>=0.1.57
langgraph-cli
langgraph-api

# Web framework dependencies
fastapi
uvicorn[standard]

# Database and vector store dependencies
chromadb>=0.4.22
langchain-chroma

# Document processing dependencies
pypdf>=4.0.0
langchain-community
pymupdf

# LLM and model dependencies
langchain-huggingface
transformers
langchain-ollama

# Image generation dependencies
diffusers
accelerate
safetensors
Pillow>=10.0.0
torchvision
torchaudio

# Development and notebook dependencies
ipykernel
jupyterlab

# Additional utilities
requests
numpy
typing-extensions